Added modules, fixed mistakes
