+++ Homeworks/Homework_4/Task3/Task3.vcxproj.filters
@@ -18,5 +18,27 @@
     <ClCompile Include="Task3.c">
       <Filter>Исходные файлы</Filter>
     </ClCompile>
+    <ClCompile Include="phoneBookOperations.c">
+      <Filter>Исходные файлы</Filter>
+    </ClCompile>
+    <ClCompile Include="phoneBookOperationsTests.c">
+      <Filter>Исходные файлы</Filter>
+    </ClCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <Text Include="phoneBook.txt">
+      <Filter>Файлы ресурсов</Filter>
+    </Text>
+    <Text Include="testReadFile.txt">
+      <Filter>Файлы ресурсов</Filter>
+    </Text>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="phoneBookOperations.h">
+      <Filter>Файлы заголовков</Filter>
+    </ClInclude>
+    <ClInclude Include="phoneBookOperationsTests.h">
+      <Filter>Файлы заголовков</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
